refactor: uts in core
