+++ core/src/miner/miner.rs
@@ -1125,7 +1125,14 @@ impl MinerService for Miner {
 
 #[cfg(test)]
 mod tests{
-    use super::*;
+//    use super::*;
+    use io::IoChannel;
+    use transaction::{SignedTransaction,PendingTransaction};
+    use super::{MinerOptions,PendingSet,Banning};
+    use spec::Spec;
+    use types::BlockNumber;
+    use std::time::Duration;
+    use std::sync::Arc;
     use aion_types::U256;
     use keychain;
     use rustc_hex::FromHex;
@@ -1135,37 +1142,6 @@ mod tests{
     use client::{BlockChainClient};
     use tests::common::{EachBlockWith, TestBlockChainClient, helpers::generate_dummy_client};
     use miner::{Miner,MinerService};
-
-    #[test]
-    fn should_prepare_block_to_seal() {
-        // given
-        let client = TestBlockChainClient::default();
-        let miner = Miner::with_spec(&Spec::new_test());
-
-        // when
-        let sealing_work = miner.map_sealing_work(&client, |_| ());
-        assert!(sealing_work.is_some(), "Expected closed block");
-    }
-    #[test]
-    fn should_still_work_after_a_couple_of_blocks() {
-        // given
-        let client = TestBlockChainClient::default();
-        let miner = Miner::with_spec(&Spec::new_test());
-
-        let res = miner.map_sealing_work(&client, |b| b.block().header().mine_hash());
-        assert!(res.is_some());
-        assert!(miner.submit_seal(&client, res.unwrap(), vec![]).is_ok());
-
-        // two more blocks mined, work requested.
-        client.add_blocks(1, EachBlockWith::Nothing);
-        miner.map_sealing_work(&client, |b| b.block().header().mine_hash());
-
-        client.add_blocks(1, EachBlockWith::Nothing);
-        miner.map_sealing_work(&client, |b| b.block().header().mine_hash());
-
-        // solution to original work submitted.
-        assert!(miner.submit_seal(&client, res.unwrap(), vec![]).is_ok());
-    }
 
     fn miner() -> Miner {
         Arc::try_unwrap(Miner::new(
@@ -1296,6 +1272,7 @@ mod tests{
 
     #[test]
     fn internal_seals_without_work() {
+        //let spec = Spec::new_instant();
         let spec = Spec::new_test();
         let mut miner = Miner::with_spec(&spec);
         miner.set_minimal_gas_price(0.into());
