+++ core/src/tests/mod.rs
@@ -22,5 +22,11 @@
 
 mod client;
 mod executive;
+mod block;
+mod miner;
+mod pod_account;
+mod pod_state;
+mod state;
+
 pub mod common;
