--- core/tests/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-description = "Core test"
-name = "acore-test"
-version = "0.1.0"
-authors = ["Aion Foundation <admin@aion.network>"]
-
-[dev-dependencies]
-acore = { path = ".." }
-acore-io = { path = "../../util/io" }
-aion-types = { path = "../../util/aion-types"}
-db = { path = "../../db/core" }
-key = { path = "../../keystore/key" }
-tempdir = "*"
